Improve the build-api compatibility script
Currently, the configure script is mostly meant for the convenience of
GNOME Continuous, but with minimal changes we can make it an appropriate
wrapper to the autotools-like build-api.

We can ensure that Meson is invoked with the location of the build
directory, as well as the source directory; we can also ensure that the
Makefile wrapper around Ninja is capable to calling Ninja with the
location of the generated build.ninja file.

Finally, we can generate a simple `check` target that calls `mesontest`
in the build directory.

These changes allow building JSON-GLib following the usual sequence:

  ./configure …
  make
  make check
  make install

While keeping all the build-related files under a build directory
(except for the generated Makefile, which can be safely ignored).