json-node: Clarify documentation regarding programmer errors
Despite my efforts in commit 011c1e424b7 (parts of which were completely
incorrect), the JsonNode documentation still wasn’t particularly helpful
about differentiating between methods which return NULL if the JsonNode
is not of the right type, and methods which emit a critical warning.

Try and improve that across the board, hopefully without saying
blatantly incorrect things this time.

Signed-off-by: Philip Withnall <withnall@endlessm.com>