ci: Add a dist job
Whenever we push a tag, we want to create a distributable archive for
json-glib; this will automate the release process.

Ideally, in the future, we'll have the ability to upload the release
archive straight to download.gnome.org, but for now store the archive
and the generated API reference as CI artefacts.