scanner: Fix crash on malformed surrogate pairs
JSON-GLib would crash when trying to decode escaped Unicode characters, if
a character of a surrogate pair was detected but it was in the wrong range, or
if the following character was not in the correct range.

Fixed by emitting an error ("Parse error: scanner: malformed surrogate pair")
before the assertion is run.

Fixes #26.