Support multiple roots in non-strict mode
Instead of replacing the root node with every new top level declaration,
and leaking the previous node in the process, we implicitly create an
array as the top level, and put every new top level under inside it.
It's not any more valid, but if you want strict conformance to the JSON
RFC, that's what the strict mode is for.

Fixes: #74