parser: Replace load_from_data()
The load_from_data() method is an example of a particularly bad pattern
in GLib-based API: a function that takes a pointer to a char array and a
length argument that can be used as a marker to tell the library to run
strlen() internally. This kind of API, neither fish nor fowl, is a
legacy from when calling strlen() was expensive; these days it's just a
way to make a mess of a pretty-straightforward API for language bindings
with a proper string type. The length argument is always going to be -1,
or it's going to be a trap, because it cannot be used as an array length
marker without turning the string into a byte array.

To replace the use case of a sub-string, we add a load_from_bytes()
method, which takes a bytes buffer; for the use case of a NUL-terminated
string, we add a load_from_string() method. This way, we have a decent
API coverage, without a weird mix between the two.