Move extern symbol macro inside the header
Instead of using the build system to define the _JSON_EXTERN macro, we
can do that straight in the version macro header, and conditionally
depend on:

- platform
- toolchain
- type of build

Additionally, we can use the appropriate compiler attribute on Windows
to ensure that intra-library symbols use the correct visibility.

This is similar to what GLib does.